Accessibility Improvements
Added the a11y plugin to storybook
Updated placeholder font color so the contrast is sufficient
Added aria-label to open/close arrow
Updated aria roles for placeholder and input
Added aria-controls for input/filterable dropdown.  Updated renderMenu function to render an empty div if it's not open. Aria rules require the aria-controlled value to be visible (even if there's nothing in it).