More accessibility improvements
Removed `aria-controls` attributes.  Did some research and it seems to only work with the JAWS reader, and not well. 
Updated aria-selected on Option and OptionGroup to reflect if an option is actually selected instead of being hard coded to false.
Added the spacebar keyCode so that it has the same behavior as tab (so it will select values as is standard with other drop downs).
Reverted logic added for aria-controls that duplicated the menu container ID.